Fix some kind of cast error on linux host
Fixing this error on specific Linux host :
```
# github.com/cloudfoundry/gosigar
../../cloudfoundry/gosigar/sigar_linux.go:93: cannot use sysinfo.Totalswap (type uint32) as type uint64 in assignment
../../cloudfoundry/gosigar/sigar_linux.go:94: cannot use sysinfo.Freeswap (type uint32) as type uint64 in assignment
```

+++ sigar_linux.go
@@ -90,8 +90,8 @@ func (self *Swap) Get() error {
 		return err
 	}
 
-	self.Total = sysinfo.Totalswap
-	self.Free = sysinfo.Freeswap
+	self.Total = uint64(sysinfo.Totalswap)
+	self.Free = uint64(sysinfo.Freeswap)
 	self.Used = self.Total - self.Free
 
 	return nil
